% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{ar5_colors}
\alias{ar5_colors}
\title{AR5 colors}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 12 rows and 5 columns.
}
\usage{
ar5_colors
}
\description{
AR5 colors
}
\keyword{datasets}
