% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{bonitet_colors}
\alias{bonitet_colors}
\title{Bonitet scale}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7 rows and 5 columns.
}
\usage{
bonitet_colors
}
\description{
Bonitet scale
}
\keyword{datasets}
