% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_boot_stat.R
\name{boot_contrast}
\alias{boot_contrast}
\alias{-.boot_stat}
\alias{/.boot_stat}
\title{Contrast functions for boot_stat object}
\usage{
boot_contrast(x, ...)

\method{-}{boot_stat}(x, ...)

\method{/}{boot_stat}(x, ...)
}
\arguments{
\item{x}{Boot object}

\item{...}{additional parameters}

\item{level}{Level to use as reference}
}
\description{
Contrast functions for boot_stat object
}
