% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_functions.R
\name{boot_mean}
\alias{boot_mean}
\alias{boot_lower}
\alias{boot_upper}
\title{Bootstrap functions for calculating mean and lower and upper bounds}
\usage{
boot_mean(x, R = 999)

boot_lower(x, limit = 0.025, R = 999)

boot_upper(x, limit = 0.975, R = 999)
}
\arguments{
\item{x}{Input vector of values to bootstrap}

\item{R}{Number of bootstrap samples. Default = 999}

\item{limit}{Confidence limit. e.g. 0.025 for lower 2.5\%, and 0.975 for upper 97.5\%}
}
\description{
Bootstrap functions for calculating mean and lower and upper bounds
}
