% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_alien_obs.R
\name{bootstrap_alien_obs}
\alias{bootstrap_alien_obs}
\title{bootstrap_alien_obs}
\usage{
bootstrap_alien_obs(
  df,
  groups = NULL,
  rolling_year_window = TRUE,
  lower_limit = 0.025,
  upper_limit = 0.975,
  Hill = TRUE,
  R = 999
)
}
\arguments{
\item{df}{dataframe (or tibble). Must be a table of presences and absences of all alien species in all localities. Typically from the table views.alien_obs and aggregated to year-locality level.}

\item{groups}{Optional grouping variables as character vector.}

\item{rolling_year_window}{Should the yearly values be calculated around a 5 year rolling window? Only applies if "year" is present in the groupings}

\item{lower_limit}{Lower limit to confidence intervals, e.g. 0.025 for lower 2.5\%}

\item{upper_limit}{Upper limit to confidence intervals, e.g. 0.975 for upper 97.5\%}

\item{Hill}{Calculate Shannon information as Hill number. Boolean}

\item{R}{Number of bootstrap samples}
}
\value{
Returns an object of class boot_stat.
}
\description{
Bootstraps alien species frequencies (number of localities with presences) for each alien species and weighs these together as summed Shannon information
}
\examples{

#'\dontrun{


}



}
