% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_value.R
\name{bootstrap_value}
\alias{bootstrap_value}
\title{Bootstrap value}
\usage{
bootstrap_value(
  df,
  value,
  groups,
  lower_limit = 0.025,
  upper_limit = 0.975,
  R = 999
)
}
\arguments{
\item{df}{dataframe (or tibble)}

\item{value}{Column to bootstrap}

\item{lower_limit}{Lower limit to confidence intervals, e.g. 0.025 for lower 2.5\%}

\item{upper_limit}{Upper limit to confidence intervals, e.g. 0.975 for upper 97.5\%}

\item{R}{Number of bootstrap samples}

\item{groupings}{Optional grouping variables as character vector.}
}
\description{
Bootstrap value
}
\examples{


\dontrun{


beetles <- obs_from_db(subset_order = "Coleoptera",
                      agg_level = "year_locality")


beetle_shannon_boot <- bootstrap_value(beetles,
                                       value = shannon_div,
                                       groups = c("year",
                                                  "region_name"))
beetle_shannon_boot

}



}
