% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_GDE.R
\name{calc_GDE}
\alias{calc_GDE}
\title{Calculate Genedic diversity evenness on a community.}
\usage{
calc_GDE(gdm, Hill = TRUE, richn_corr = TRUE)
}
\arguments{
\item{Hill}{Return as first order Hill number. Boolean}

\item{richn_corr}{Correct for species richness, dividing by number of species. Boolean}

\item{community}{Community.}
}
\value{
Shannon diversity.
}
\description{
Experimental!
}
\examples{


\dontrun{
   beetles_2021 <- get_observations(subset_orders = "Coleoptera",
   subset_year = 2021,
   agg_level = "none")


  shannon_beetles_2021 <- beetles_2021 \%>\%
  collect() \%>\%
  group_by(locality) \%>\%
  select(locality,
        no_asv_per_species) \%>\%
  summarise(shannon_div = calc_GDE(species_latin))


}
}
