% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_shannon.R
\name{calc_shannon}
\alias{calc_shannon}
\title{Calculate Shannon index on a community.}
\usage{
calc_shannon(
  community,
  no_asv_per_species,
  augment_community = TRUE,
  only_sum_freq = FALSE,
  Hill = TRUE
)
}
\arguments{
\item{community}{Community.}

\item{no_asv_per_species}{Number of genetic variants per species.}

\item{augment_community}{Build up community by from species and number of occurences (TRUE), or use raw occurrences (FALSE)}

\item{only_sum_freq}{Only sum the frequencies together, instead of calculating Shannon information.}

\item{Hill}{Return as first order Hill number. Boolean}
}
\value{
Shannon diversity.
}
\description{
Calculate Shannon index on a community.
}
\examples{


\dontrun{
   beetles_2021 <- get_observations(subset_orders = "Coleoptera",
   subset_year = 2021,
   agg_level = "none")


  shannon_beetles_2021 <- beetles_2021 \%>\%
  collect() \%>\%
  group_by(locality) \%>\%
  select(locality,
        species_latin) \%>\%
  summarise(shannon_div = calc_shannon(species_latin))


}
}
