% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCon.R
\name{checkCon}
\alias{checkCon}
\title{Check for an open connection to the database.}
\usage{
checkCon()
}
\value{
An error message if a connection called con is missing
}
\description{
Of internal use
}
\examples{

\dontrun{
checkCon()

}

}
