% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clim_trend_plot.R
\encoding{UTF-8}
\name{clim_trend_plot}
\alias{clim_trend_plot}
\title{clim_trend_plot}
\usage{
clim_trend_plot(
  region = c("Trøndelag", "Østlandet", "Sørlandet", "Nord-Norge", "Vestlandet"),
  dataset = c("NasIns", "TidVar", "ØkoTrond", "HulEik"),
  from_year = 2010,
  to_year = 2023,
  from_month = 6,
  to_month = 7,
  subset_loc_in_last_year = TRUE
)
}
\arguments{
\item{region}{Which region to plot? Currently available: c("Trøndelag", "Østlandet", "Sørlandet", "Nord-Norge", "Vestlandet").}

\item{dataset}{Which dataset to use? Is used to select the localities within regions.}

\item{from_year}{Plot data from what year= Integer.}

\item{to_year}{Plot data to what year? Integer.}

\item{from_month}{Summarise data from what month?. Integer}

\item{to_month}{Summarise data to what month? Integer}

\item{subset_loc_in_last_year}{Should we only plot data for the localities visited in the last year? Boolean.}
}
\value{
A plot (a grob containing two ggplots)
}
\description{
Summarises mean values of the weather (temperature and precipitation) during chosen months for the localities within a chosen region.
}
\examples{
\dontrun{

clim_trend_plot(
  region = "Trøndelag",
  from_year = 2010,
  to_year = 2022,
  from_month = 6,
  to_month = 7
)
}

}
