% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_dist_to_comm_mat.R
\name{combine_dist_to_comm_mat}
\alias{combine_dist_to_comm_mat}
\title{combine_dist_to_comm_mat}
\usage{
combine_dist_to_comm_mat(comm_mat, region_name, habitat_type)
}
\arguments{
\item{habitat_type}{}
}
\description{
combine_dist_to_comm_mat
}
