% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_matrix_from_db.R
\name{community_matrix_from_db}
\alias{community_matrix_from_db}
\title{community_matrix_from_db}
\usage{
community_matrix_from_db(
  limit = NULL,
  id_type = c("metabarcoding"),
  trap_type = c("MF", "VF", "All", NULL),
  dataset = c("NasIns"),
  subset_orders = NULL,
  subset_families = NULL,
  subset_species = NULL,
  subset_habitat = NULL,
  subset_region = c(NULL, "Østlandet", "Trøndelag"),
  exclude_singletons = F,
  transposed_matrix = F,
  as_tibble = F
)
}
\arguments{
\item{as_tibble}{}
}
\description{
community_matrix_from_db
}
