% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_database.R
\name{connect_to_database}
\alias{connect_to_database}
\title{connect_to_database}
\usage{
connect_to_database(
  username,
  password,
  host = "ninradardata01.nina.no",
  dbname = "insect_monitoring",
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Best practice is to not store your username and password in any scripts. See example for a way to script retrieval of credentials
}
\examples{


Connect to insect monitoring database while storing your credentials in a password file

dontrun{

  source("~/.rpgpass")

  connect_to_database(
     username = username,
     password = password
  )

  rm(list = c("username", "password"))
}


Connect to another database

connect_to_database(
  username = "postgjest",
  password = "gjestpost",,
  host = "gisdata-db.nina.no",
  dbname = "gisdata"
)


}
