% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_asv_loc.R
\name{get_asv_loc}
\alias{get_asv_loc}
\title{get_asv_loc}
\usage{
get_asv_loc(species = NULL, dataset = NULL, subset_years = NULL)
}
\arguments{
\item{species}{What species to fetch records for. Character string.}

\item{dataset}{Optionally subset to a (character vector of) project(s).}

\item{subset_years}{Optionally subset vector of years}
}
\value{
A tibble with asv data.
}
\description{
get_asv_loc
}
\examples{
\dontrun{

anc_bad_asv <- get_asv_loc("Ancylis badiana")

}


}
