% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_biomass.R
\name{get_biomass}
\alias{get_biomass}
\title{get_biomass Get biomass data from the database}
\usage{
get_biomass(
  limit = NULL,
  trap_type = "MF",
  subset_region = c(NULL, "Østlandet", "Trøndelag", "Sørlandet"),
  dataset = "NasIns",
  agg_level = "year_locality",
  as_tibble = F
)
}
\arguments{
\item{limit}{Optional row limit on output (for testing)}

\item{trap_type}{Which trap types to fetch data for. "All" (defult), "MF", or "VF".}

\item{subset_region}{Optional subset of region. Currently allowing "Østlandet", "Trøndelag", "Sørlandet"}

\item{dataset}{Which dataset to fetch data for. Default "NasIns".}

\item{agg_level}{Aggregate level of data. "year_locality" (default), "locality_sampling", "total".}

\item{as_tibble}{Return as tibble? Boolean}
}
\value{
A tibble of sample weights
}
\description{
get_biomass Get biomass data from the database
}
\examples{

\dontrun{

connect_to_insect_db()
biomass <- get_biomass(agg_level = "year_locality",
trap_type = "MF")

}

}
