% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_climate_data.R
\name{get_climate_data}
\alias{get_climate_data}
\title{get_climate_data}
\usage{
get_climate_data(locality = NULL)
}
\arguments{
\item{locality}{A character vector of the locality name, e.g. "Semi-nat_01", or c("Semi-nat_32", "Semi-nat_33", ...)}
}
\value{
A tibble of daily climate data from 1957.
}
\description{
Fetch climatic records for a particular locality. Data includes daily mean temperature, total precipitation, and total snow depth. Data stretches from 1957 up until the latest import to the database.
}
\examples{
\dontrun{
climate_data <- get_climate_data("Semi-nat_01")
}

}
