% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_forest_type.R
\name{get_forest_type}
\alias{get_forest_type}
\title{get_forest_type Get the forest types from a buffer around the sampling localities}
\usage{
get_forest_type(
  limit = NULL,
  dataset = c("NasIns"),
  stat_grouping = c("treslag", "bonitet"),
  as_tibble = F
)
}
\arguments{
\item{limit}{Optional row limit of returned data (for testing )}

\item{dataset}{Dataset to get data from. Defaults to "NasIns"}

\item{stat_grouping}{Group data into either "treslag" or "bonitet"}

\item{as_tibble}{Return as tibble? Boolean}
}
\value{
Returns a tibble of forest data from the database.
}
\description{
get_forest_type Get the forest types from a buffer around the sampling localities
}
\examples{
connect_to_insect_db()
forest_data <- get_forest_type()


}
