% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_id_types.R
\name{get_id_types}
\alias{get_id_types}
\title{Fetch table of identification types in the database.}
\usage{
get_id_types(include_project_years = TRUE)
}
\arguments{
\item{include_project_years}{Include list of which projects and years each method was used? Boolean, default = TRUE.}
}
\value{
A tibble of the identification names (optionally with the projects they where used in)
}
\description{
Used to lookup what "id_types" are available to filter on in e.g. get_observations(). The names in the "identification_types" column can be us
}
\examples{

\dontrun{

get_id_types()

tidvar_2019_legs <- get_observations(id_type = "legs_col_hem_lep", dataset = "TidVar")

}


}
