% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dowload_licor_logger_data.R
\name{get_licor_logger_data}
\alias{get_licor_logger_data}
\title{get_licor_logger_data
Downloads logger data from the licor.cloud platform}
\usage{
get_licor_logger_data(
  loggers = NULL,
  start_date = NULL,
  end_date = NULL,
  token = NULL
)
}
\arguments{
\item{loggers}{String of logger id's to fetch data for}

\item{start_date}{Start date of data (YYYY-MM-DD format)}

\item{end_date}{End date of data (YYYY-MM-DD format)}

\item{token}{Autorization token (long string, created at https://licor.cloud)}
}
\value{
A tibble of logger data
}
\description{
get_licor_logger_data
Downloads logger data from the licor.cloud platform
}
\examples{

\dontrun{

logger_data <- download_licor_logger_data(loggers = 20878881,
                                          start_date = "2024-06-01",
                                          end_date = "2024-06-30",
                                          token = "your_token")


}

}
