% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_localities.R
\name{get_localities}
\alias{get_localities}
\title{get_localities}
\usage{
get_localities(
  dataset = c("NasIns", "OkoTrond", "TidVar", "Nerlandsøya", "All"),
  as_sf = TRUE,
  habitat_type = c("All", "Semi-nat", "Forest", "Tid_man", "Tid_aut", "Wetland")
)
}
\arguments{
\item{dataset}{Which dataset to fetch ("NasIns", "OkoTrond", "TidVar", "Nerlandsøya", "All") . Default to "NasIns" for the national insect monitoring data}

\item{as_sf}{Return an sf object? Boolean.}

\item{habitat_type}{Optional habitat type to subset}
}
\value{
An sf or tibble with the location info
}
\description{
Get the localities from the database, either as an sf object or as a tibble
}
\examples{
\dontrun{
connect_to_insect_db()
loc <- get_localities()
ggplot(loc) +
geom_sf(aes(fill = habitat_type))

}

}
