% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_loggerdata.R
\name{get_logger_data}
\alias{get_logger_data}
\title{get_logger_data Get temperature and light logger data from the database}
\usage{
get_logger_data(
  limit = NULL,
  dataset = c("NorIns"),
  agg_level = c("year_locality", "locality_sampling"),
  as_tibble = F
)
}
\arguments{
\item{limit}{Optional row number limit (for testing)}

\item{dataset}{Which dataset to retreive data from? Default to "NorIns".,}

\item{agg_level}{Aggregation level of data. Character, either "year_locality" (default) or "locality_sampling"}

\item{as_tibble}{Return results as tibble? Boolean.}
}
\value{
Tibble or dataframe of logger data.
}
\description{
get_logger_data Get temperature and light logger data from the database
}
\examples{
\dontrun{
locality_sampling_loggerdata <- get_logger_data(
  dataset = "NorIns",
  agg_level = "locality_sampling"
)
}

}
