% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_observations.R
\name{get_observations}
\alias{get_observations}
\title{get_observations Get insect observation data from the database}
\usage{
get_observations(
  id_type = c("metabarcoding"),
  subset_orders = NULL,
  subset_families = NULL,
  subset_genus = NULL,
  subset_species = NULL,
  subset_year = NULL,
  subset_region = NULL,
  subset_habitat = NULL,
  trap_type = "All",
  limit = NULL,
  dataset = "NasIns",
  agg_level = "year_locality",
  digits = 2,
  as_tibble = F
)
}
\arguments{
\item{id_type}{Type of identification type. Defaults to metabarcoding data}

\item{subset_orders}{Optional subset of order}

\item{subset_families}{Optional subset of families}

\item{subset_genus}{Optional subset of genus}

\item{subset_species}{Optional subset of species}

\item{subset_year}{Optional subset of year}

\item{subset_region}{Optional subset of region}

\item{subset_habitat}{Optional subset of habitat type ("Forest" or "Semi-nat")}

\item{trap_type}{Optional subset of trap type}

\item{limit}{Optional limit the output to number of rows (for testing)}

\item{dataset}{Choose the dataset to fetch data from. Default "NasIns" for national insect monitoring data}

\item{agg_level}{Aggregation level of data. "year_locality", "region_habitat", "region_habitat_year", "total". Default to year_locality}

\item{digits}{Number of digits to round shannon diversity and mean ASV counts to. (defaults to 2)}

\item{as_tibble}{Coerce output to class tibble}
}
\value{
A tibble of insect observations from the database
}
\description{
get_observations Get insect observation data from the database
}
\examples{

dontrun{

  source("~/.rpgpass")

  connect_to_database(
     username = username,
     password = password
  )

  rm(list = c("username", "password"))


  beetles_2022 <- get_observations(subset_orders = "Coleoptera",
                              agg_level = "year_locality")

}



}
