% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_phenology.R
\name{get_phenology}
\alias{get_phenology}
\title{get_phenology Get insect phenology data from the database}
\usage{
get_phenology(
  taxonomic_level = NULL,
  id_type = c("metabarcoding"),
  subset_year = NULL,
  subset_region = NULL,
  subset_habitat = NULL,
  subset_order = NULL,
  trap_type = "All",
  limit = NULL,
  dataset = "NorIns",
  digits = 2,
  return_tibble = F
)
}
\arguments{
\item{taxonomic_level}{taxonomic level for the comparison}

\item{id_type}{Type of identification. Defaults to metabarcoding data}

\item{subset_year}{Optional subset of year}

\item{subset_region}{Optional subset of region}

\item{subset_habitat}{Optional subset of habitat type ("Forest" or "Semi-nat")}

\item{trap_type}{Optional subset of trap type}

\item{limit}{Optional limit the output to number of rows (for testing)}

\item{dataset}{Choose the dataset to fetch data from. Default "NasIns" for national insect monitoring data}

\item{digits}{Number of digits to round shannon diversity and mean ASV counts to. (defaults to 2)}

\item{as_tibble}{Coerce output to class tibble}
}
\value{
A tibble of insect observations from the database
}
\description{
This provides a summary of insect catches, divided into taxa of a chosen level over the season, typically to be plotted by plot_phenology().
}
\examples{
\dontrun{

  connect_to_insect_db()


  beetles <- get_observations(subset_orders = "Coleoptera",
                              agg_level = "year_locality")

}



}
