% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pollinators.R
\name{get_pollinators}
\alias{get_pollinators}
\title{Get list of pollinator families (e.g. for fetching with obs_from_db)}
\usage{
get_pollinators()
}
\value{
A tibble of norwegian an latin names of pollinator families.
}
\description{
Get list of pollinator families (e.g. for fetching with obs_from_db)
}
\examples{

\dontrun{
pollinators <- get_pollinators()

}


}
