% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_projects.R
\name{get_projects}
\alias{get_projects}
\title{Get list of all projects in the insect_monitoring DB}
\usage{
get_projects()
}
\value{
a tibble with the projects, where the names 'projects_short_name' is used for subsetting data in other functions
}
\description{
Get list of all projects in the insect_monitoring DB
}
\examples{
\dontrun{
get_projects()

}

}
