% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample_info.R
\name{get_sample_info}
\alias{get_sample_info}
\title{Get basic sample info (locality, date and trap type)}
\usage{
get_sample_info(dataset = c("NorIns"))
}
\arguments{
\item{dataset}{Which dataset to fetch data for.}
}
\value{
A tibble with sample info
}
\description{
Get basic sample info (locality, date and trap type)
}
\examples{

\dontrun{
get_sample_info(dataset = "NorIns")

}

}
