% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_year_region_stats.R
\name{get_year_region_stats}
\alias{get_year_region_stats}
\title{get_year_region_stats}
\usage{
get_year_region_stats(project_short_name = "NorIns", last_year = 2024)
}
\arguments{
\item{project_short_name}{The project to fetch data for (Default: "NorIns")}

\item{last_year}{The last year of the data to fetch info for}
}
\value{
A tibble of project year data
}
\description{
Utility function, primarily for the plot_year_locality_sum plot function in the external shiny app
}
\examples{

\dontrun{

to_project_year_plot <- get_year_region_stats(last_year = 2025)

}

}
