% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loggerdata_from_db.R
\name{loggerdata_from_db}
\alias{loggerdata_from_db}
\title{logger_data_from_db Get temperature and light logger data from the database}
\usage{
loggerdata_from_db(
  limit = NULL,
  dataset = c("NasIns"),
  agg_level = c("year_locality", "locality_sampling"),
  as_tibble = F
)
}
\arguments{
\item{as_tibble}{}
}
\description{
logger_data_from_db Get temperature and light logger data from the database
}
