% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{longerHobo2202}
\alias{longerHobo2202}
\title{Reformat hobo data from the 2202 logger types}
\usage{
longerHobo2202(inputFile, guess_max = 10000)
}
\description{
Reformat hobo data from the 2202 logger types
}
