% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{longerHobo2301}
\alias{longerHobo2301}
\title{Reformat hobo data from the 2301A logger types}
\usage{
longerHobo2301(inputFile, guess_max = 10000, ...)
}
\description{
Reformat hobo data from the 2301A logger types
}
