% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_plot_boot_stat.R
\name{map_plot}
\alias{map_plot}
\title{map_plot Plots bootstrap values from boot_stat object on map}
\usage{
map_plot(
  x,
  whole_country = FALSE,
  alpha_from_sd = FALSE,
  alpha_range = c(0.3, 0.9),
  palette = "blue-orange",
  ...
)
}
\arguments{
\item{x}{A boot_stat object}

\item{whole_country}{Plot whole country (and not only regions with values). Boolean}

\item{alpha_from_sd}{Boolean, compute alpha (transparancy) values from sd?}

\item{alpha_range}{Range of alpha values}

\item{...}{Additional parameters passed to scale_fill_nina. e.g. palette}
}
\value{
A ggplot
}
\description{
map_plot Plots bootstrap values from boot_stat object on map
}
\examples{


\dontrun{


connect_to_insect_db()

beetles <- get_observations(subset_orders = "Coleoptera",
                      agg_level = "year_locality")


beetle_asv_div_boot <- bootstrap_value(beetles,
                                      value = mean_asv_per_species,
                                      groups = c("year",
                                                 "region_name"))


map_plot(beetle_asv_div_boot,
         palette = "blue-orange")

}
}
