% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_from_db.R
\name{obs_from_db}
\alias{obs_from_db}
\title{-obs_from_db Get insect observation data from the database}
\usage{
obs_from_db(
  id_type = c("metabarcoding"),
  subset_orders = NULL,
  subset_families = NULL,
  subset_genus = NULL,
  subset_species = NULL,
  subset_year = NULL,
  subset_region = c(NULL, "Østlandet", "Trøndelag"),
  trap_type = c("All", "MF", "VF"),
  limit = NULL,
  dataset = c("NasIns"),
  agg_level = c("year_locality", "locality_sampling", "region_habitat",
    "region_habitat_year", "total"),
  as_tibble = F
)
}
\arguments{
\item{id_type}{Type of identification type. Defaults to metabarcoding data}

\item{subset_orders}{Optional subset of order}

\item{as_tibble}{}
}
\description{
-obs_from_db Get insect observation data from the database
}
