% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phen_peak_plot.R
\name{phen_peak_plot}
\alias{phen_peak_plot}
\title{Phen_peak_plot
Plot the peaks of a sinusiod phenology model output.}
\usage{
phen_peak_plot(obj, mark_peaks = TRUE, pred_time = 1:365)
}
\arguments{
\item{obj}{a gllvm sinusoid model.}

\item{mark_peaks}{Mark peaks? boolean}

\item{pred_time}{x-axis length, prediction interval. Defaults to 365 days.}
}
\value{
a ggplot object
}
\description{
Phen_peak_plot
Plot the peaks of a sinusiod phenology model output.
}
\examples{

\dontrun{

phen_peak_plot(gllvm_phen_model)

}

}
