% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ano_herb_sum.R
\name{plot_ano_herb_sum}
\alias{plot_ano_herb_sum}
\title{Summary plot of ano herb records}
\usage{
plot_ano_herb_sum(subset_year = NULL)
}
\arguments{
\item{subset_year}{Optional subset of year (e.g. 2020:2024 or character vector)}
}
\value{
A ggplot
}
\description{
Convenience function to clear up some code in reports
}
\examples{
\dontrun{

plot_ano_herb_sum()
}

}
