% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_asv.R
\name{plot_asv}
\alias{plot_asv}
\title{plot_asv}
\usage{
plot_asv(
  species = NULL,
  background = "ortofoto/hele_norge.png",
  pie_scale = 0.5,
  size = 0.1,
  caption = TRUE,
  title = NULL,
  scale_to_sum_reads = TRUE,
  ...
)
}
\arguments{
\item{species}{Character string of latin species name to plot}

\item{background}{relative path to a background image.}

\item{pie_scale}{scaling factor relative to map size (how large should the pies be) 0-1}

\item{size}{another (?) sizing}

\item{caption}{Use species name as caption? Boolean}

\item{title}{Pptional title. Null or character.}

\item{scale_to_sum_reads}{Should slices be scaled to the sum of the reads? Boolean}

\item{...}{}
}
\value{
a ggplot object
}
\description{
plot_asv
}
\examples{
\dontrun{

plot_asv("Erebia ligea") #'
}

}
