% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_beta_part.R
\name{plot_beta_part}
\alias{plot_beta_part}
\title{plot_beta_part}
\usage{
plot_beta_part(input)
}
\arguments{
\item{input}{Tibble or dataframe of betapartition + locality distance from the combine_dist_to_comm_mat function}
}
\value{
Returns a plot of the beta diversity partition of the overall diversity
}
\description{
plot_beta_part
}
\examples{


#' skog_ost_comm_mat <- community_matrix_from_db(trap_type = "MF",
dataset = "NasIns",
subset_habitat = "Forest",
subset_region = "Østlandet",
as_tibble = T,
transposed_matrix = F
) \%>\%
  select(-c(year, locality))

skog_ost_dist_beta <- combine_dist_to_comm_mat(comm_mat = skog_ost_comm_mat,
region_name = "('Østlandet')",
habitat_type = "Forest")

plot_beta_part(skog_ost_dist_beta)


}
