% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_climate_comparison.R
\name{plot_climate_comparison}
\alias{plot_climate_comparison}
\title{plot_climate_comparison}
\usage{
plot_climate_comparison(
  climate_data = NULL,
  variable = c("temperature", "precipitation", "snow_depth"),
  focus_year = "latest",
  clip_to_1990 = TRUE,
  y_high_limit = 60,
  y_low_limit = -30,
  language = c("Norwegian", "English"),
  main_title = TRUE,
  rolling_mean = FALSE
)
}
\arguments{
\item{climate_data}{A tibble of climate data (typically from get_climate_data())}

\item{variable}{Which variable to choose. "temperature"(default), "precipitation", or "snow_depth"}

\item{focus_year}{Focus on "latest" year (default) or optional year as numerical value (e.g. 2021)}

\item{clip_to_1990}{Logical, should historical records be clipped to >1990.}

\item{y_high_limit}{The y-axis limit of the plot.}

\item{language}{Figure text in "Norwegian" or "English"}

\item{main_title}{Print plot title? Boolean.}

\item{rolling_mean}{Plot rolling mean of 5 days? Suitable for precipitation}
}
\value{
A ggplot
}
\description{
Plot the normal climate range and a chosen year as a comparison. Code adapted from Brad Boehmke and Scott Ogletree.
}
\examples{

\dontrun{

connect_to_insect_db()
get_climate_data("Semi-nat_01") \%>\%
plot_climate_comparison()

}

}
