% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_country_map.R
\name{plot_country_map}
\alias{plot_country_map}
\title{plot_country_map}
\usage{
plot_country_map(map, style = "regions")
}
\arguments{
\item{map}{Map input, typically form Norimon::get_map()}

\item{style}{What style to plot, currently only "regions" implemented}
}
\value{
A ggplot2 object
}
\description{
plot_country_map
}
\examples{
\dontrun{

nor <- get_map()

plot_country_map(nor)

}

}
