% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rast_list}
\alias{rast_list}
\title{Function to make list of rasters out of many localities. To be used with fasterize}
\usage{
rast_list(input)
}
\description{
Function to make list of rasters out of many localities. To be used with fasterize
}
