% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skogtype_from_db.R
\name{skogtype_from_db}
\alias{skogtype_from_db}
\title{skogtype_from_db Get the forest types from a buffer around the sampling localities}
\usage{
skogtype_from_db(
  limit = NULL,
  dataset = c("NasIns"),
  stat_grouping = c("treslag", "bonitet"),
  as_tibble = F
)
}
\arguments{
\item{as_tibble}{}
}
\description{
skogtype_from_db Get the forest types from a buffer around the sampling localities
}
