% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{treslag_colors}
\alias{treslag_colors}
\title{Treslag scale}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 5 columns.
}
\usage{
treslag_colors
}
\description{
Treslag scale
}
\keyword{datasets}
