% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_plot_boot_stat.R
\name{ts_plot}
\alias{ts_plot}
\title{ts_plot Plot a time series plot of indicator values}
\usage{
ts_plot(x, palette = "blue-orange", ...)
}
\arguments{
\item{x}{A boot_stat object}

\item{palette}{The palette to be used in scale_fill_nina and scale_color_nina}

\item{...}{Additional parameters passed to scale_fill_nina. e.g. palette}
}
\value{
a ggplot
}
\description{
ts_plot Plot a time series plot of indicator values
}
