% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_from_db.R
\name{weights_from_db}
\alias{weights_from_db}
\title{weights_from_db Get biomass data from the database}
\usage{
weights_from_db(
  limit = NULL,
  trap_type = c("ALL", "MF", "VF"),
  subset_region = c(NULL, "Østlandet", "Trøndelag"),
  dataset = c("NasIns"),
  agg_level = c("year_locality", "locality_sampling", "total"),
  as_tibble = F
)
}
\arguments{
\item{as_tibble}{}
}
\description{
weights_from_db Get biomass data from the database
}
