% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_shannon.R
\name{calc_shannon}
\alias{calc_shannon}
\title{Calculate Shannon index on a community}
\usage{
calc_shannon(community)
}
\arguments{
\item{community}{Community.}
}
\value{
Shannon diversity.
}
\description{
Calculate Shannon index on a community
}
