% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_shannon.R
\name{calc_shannon}
\alias{calc_shannon}
\title{Calculate Shannon index on a community.}
\usage{
calc_shannon(community)
}
\arguments{
\item{community}{Community.}
}
\value{
Shannon diversity.
}
\description{
Calculate Shannon index on a community.
}
\examples{


\dontrun{
   beetles_2021 <- get_observations(subset_orders = "Coleoptera",
   subset_year = 2021,
   agg_level = "none")


  shannon_beetles_2021 <- beetles_2021 \%>\%
  collect() \%>\%
  group_by(locality) \%>\%
  select(locality,
        species_latin) \%>\%
  summarise(shannon_div = calc_shannon(species_latin))


}
}
