% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_correlation.R
\name{chart_correlation}
\alias{chart_correlation}
\title{chart.Correlation2}
\usage{
chart_correlation(
  R,
  histogram = TRUE,
  method = c("pearson", "kendall", "spearman"),
  color = "red",
  ...
)
}
\arguments{
\item{...}{}
}
\description{
chart.Correlation2
}
