% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_dist_to_comm_mat.R
\name{combine_dist_to_comm_mat}
\alias{combine_dist_to_comm_mat}
\title{combine_dist_to_comm_mat Partitions diversity components and merges it with distance matrix of localities. This is not pretty, but does the job for a report analysis}
\usage{
combine_dist_to_comm_mat(comm_mat, region_name, habitat_type)
}
\arguments{
\item{comm_mat}{Community matrix}

\item{region_name}{Region name to subset localities from}

\item{habitat_type}{Habitat type to subset localities from}
}
\value{
A tibble of beta diversities and distances between localities
}
\description{
combine_dist_to_comm_mat Partitions diversity components and merges it with distance matrix of localities. This is not pretty, but does the job for a report analysis
}
\examples{

\dontrun{

skog_ost_comm_mat <- community_matrix_from_db(trap_type = "MF",
dataset = "NasIns",
subset_habitat = "Forest",
subset_region = "Østlandet",
as_tibble = T,
transposed_matrix = F
) \%>\%
  select(-c(year, locality))

skog_ost_dist_beta <- combine_dist_to_comm_mat(comm_mat = skog_ost_comm_mat,
region_name = "('Østlandet')",
habitat_type = "Forest")

plot_beta_part(skog_ost_dist_beta)

}

}
