% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_insect_db.R
\name{connect_to_insect_db}
\alias{connect_to_insect_db}
\title{connect_to_insect_db}
\usage{
connect_to_insect_db(
  myusername = username,
  mypassword = password,
  host = "ninradardata01.nina.no",
  dbname = "insect_monitoring",
  ...
)
}
\arguments{
\item{myusername}{Username as character. Will look for a variable called username as default.}

\item{mypassword}{Password as character. Will look for a variable called password as default.}

\item{host}{Hostname of database}

\item{dbname}{Database name}

\item{...}{}
}
\value{
A PostgreS connection to a database called "con"
}
\description{
Best practice is to not store your username and password in any scripts. See example for a way to script retrieval of credentials
}
\examples{


Connect to insect monitoring database while storing your credentials in a password file

dontrun{

  source("~/.rpgpass")

  connect_to_insect_db(
     username = username,
     password = password
  )

  rm(list = c("username", "password"))
}


Connect to another database

connect_to_insect_db(
  username = "postgjest",
  password = "gjestpost",,
  host = "gisdata-db.nina.no",
  dbname = "gisdata"
)


}
