% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_insect_db.R
\name{connect_to_insect_db}
\alias{connect_to_insect_db}
\title{connect_to_insect_db}
\usage{
connect_to_insect_db(
  host = "T2lippgsql04.nina.no",
  dbname = "insect_monitoring",
  ...
)
}
\arguments{
\item{host}{Hostname of database}

\item{dbname}{Database name}

\item{...}{additional parameter to DBI:dbConnect, such as username and password, if not provided in ~/.pgpass}
}
\value{
A PostgreSQL connection to a database called "con"
}
\description{
Best practice is to not store your username and password in any scripts. See example for a way to script retrieval of credentials
}
\examples{


#Connect to insect monitoring database while storing your credentials in a password file (~/.pgpass)

\dontrun{


  connect_to_insect_db()



Connect to another database

connect_to_insect_db(
  username = "postgjest",
  password = "gjestpost",,
  host = "gisdata-db.nina.no",
  dbname = "gisdata"
)

}


}
