% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_map.R
\name{get_map}
\alias{get_map}
\title{Get map of terrestrial norway}
\usage{
get_map(region_subset = NULL)
}
\arguments{
\item{region_subset}{Optional character vector of regions to subset.}
}
\value{
An sf-object of the terrestrial landmass of Norway, contingent on region subset.
}
\description{
Get map of terrestrial norway
}
\examples{


\dontrun{

   norway <- get_map()

norway <- norway \%>\%
 left_join(beetle_shannon_boot,
           by = c("region" = "region_name")) \%>\%
 replace_na(list(year =  2021))


norway \%>\%
 filter(year == 2021) \%>\%
 ggplot() +
 geom_sf(aes(fill = boot_mean)) +
 scale_fill_nina(name = "Beetle Shannon diversity",
                 discrete = FALSE)


}



}
