% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_observations.R
\name{get_observations}
\alias{get_observations}
\title{get_observations Get insect observation data from the database}
\usage{
get_observations(
  dataset = "NorIns",
  id_type = NULL,
  trap_type = "All",
  subset_orders = NULL,
  subset_families = NULL,
  subset_genus = NULL,
  subset_species = NULL,
  subset_year = NULL,
  subset_region = NULL,
  subset_habitat = NULL,
  limit = NULL,
  agg_level = "year_locality",
  Hill = TRUE,
  richn_corr = TRUE,
  digits = 2,
  as_tibble = F
)
}
\arguments{
\item{dataset}{Choose the dataset to fetch data from. Default "NorIns" for national insect monitoring data}

\item{id_type}{Optional filtering on identification/sampling technique, check get_id_types() for available options. Defaults to NULL with no filtering.}

\item{trap_type}{Optional subset of trap type}

\item{subset_orders}{Optional subset of order}

\item{subset_families}{Optional subset of families}

\item{subset_genus}{Optional subset of genus}

\item{subset_species}{Optional subset of species}

\item{subset_year}{Optional subset of year}

\item{subset_region}{Optional subset of region}

\item{subset_habitat}{Optional subset of habitat type ("Forest" or "Semi-nat")}

\item{limit}{Optional limit the output to number of rows (for testing)}

\item{agg_level}{Aggregation level of data. "year_locality", "region_habitat", "region_habitat_year", "locality_sampling", "total". Default to year_locality}

\item{Hill}{Calculate shannon diversity as Hill number (exp(Shann_div)). Boolean}

\item{richn_corr}{Correct GDE calculation by number of species (doi: https://doi.org/10.1101/2022.02.09.479762). Boolean.}

\item{digits}{Number of digits to round shannon diversity and mean ASV counts to. (defaults to 2)}

\item{as_tibble}{Coerce output to class tibble}
}
\value{
A tibble of insect observations from the database
}
\description{
get_observations Get insect observation data from the database
}
\examples{
\dontrun{

  connect_to_insect_db()

  beetles <- get_observations(subset_orders = "Coleoptera",
                              agg_level = "year_locality")

}

}
