% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locality_ar5_plot.R
\name{locality_ar5_plot}
\alias{locality_ar5_plot}
\title{locality_ar5_plot}
\usage{
locality_ar5_plot(
  locality,
  ano_circles = T,
  title = T,
  legend = T,
  strip = F,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
locality_ar5_plot
}
