% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longerHobo2202.R
\name{longerHobo2202}
\alias{longerHobo2202}
\title{longerHobo2202
Convenience function to read in a raw file of hobo logger MX2202 data and rearrange the columns to make is usable}
\usage{
longerHobo2202(
  inputFile,
  guess_max = 10000,
  delim = ";",
  date_format = "\%d/\%m/\%y \%H:\%M:\%S \%z",
  ...
)
}
\arguments{
\item{inputFile}{export file of raw logger data from HoboLink.com.}

\item{guess_max}{optional number of rows to guess the data format from. May have to increase a lot.}

\item{delim}{Character string of delimiter in source data. Defaults to ";".}

\item{date_format}{Character vector of date format. Be aware of the day, month, year ordering, and any trailing \%z for timezone info.}

\item{...}{additional parameters passed to read_csv}
}
\value{
A tibble of well formatted logger data from the MX2301A type logger
}
\description{
longerHobo2202
Convenience function to read in a raw file of hobo logger MX2202 data and rearrange the columns to make is usable
}
\examples{
\dontrun{

mx2202_data <- longerHobo2202(inputFile = "../rawData/temp_logger/
NasIns_2021_all_MX2202_2021_11_15_10_58_26_CET_1.csv")
}


}
