% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longerHobo2301.R
\name{longerHobo2301}
\alias{longerHobo2301}
\title{longerHobo2301}
\usage{
longerHobo2301(
  inputFile,
  guess_max = 10000,
  delim = ";",
  date_format = "\%d/\%m/\%y \%H:\%M:\%S \%z",
  ...
)
}
\arguments{
\item{inputFile}{export file of raw logger data from HoboLink.com}

\item{guess_max}{optional number of rows to guess the data format from. May have to increase a lot}

\item{delim}{Character string of delimiter in source data. Defaults to ";".}

\item{date_format}{Character vector of date format. Be aware of the day, month, year ordering, and any trailing \%z for timezone info.}

\item{...}{additional arguments passed to read_csv}
}
\value{
A tibble of well formatted logger data from the MX2301A type logger
}
\description{
Convenience function to read in a raw file of hobo logger MX2301A data and rearrange the columns to make is usable
}
\examples{
\dontrun{

mx2301A_data <- longerHobo2301(inputFile = "../rawData/temp_logger/
NasIns_2021_all_MX2301A_2021_11_15_11_00_05_CET_1.csv")
}

}
