% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_from_db.R
\name{obs_from_db}
\alias{obs_from_db}
\title{obs_from_db Get insect observation data from the database}
\usage{
obs_from_db(
  id_type = c("metabarcoding"),
  subset_orders = NULL,
  subset_families = NULL,
  subset_genus = NULL,
  subset_species = NULL,
  subset_year = NULL,
  subset_region = c(NULL, "Østlandet", "Trøndelag"),
  trap_type = c("All", "MF", "VF"),
  limit = NULL,
  dataset = c("NasIns"),
  agg_level = c("year_locality", "locality_sampling", "region_habitat",
    "region_habitat_year", "total"),
  as_tibble = F
)
}
\arguments{
\item{id_type}{Type of identification type. Defaults to metabarcoding data}

\item{subset_orders}{Optional subset of order}

\item{subset_families}{Optional subset of families}

\item{subset_genus}{Optional subset of genus}

\item{subset_species}{Optional subset of species}

\item{subset_year}{Optional subset of year}

\item{subset_region}{Optional subset of region}

\item{trap_type}{Optional subset of trap type}

\item{limit}{Optional limit the output to number of rows (for testing)}

\item{dataset}{Choose the dataset to fetch data from. Default "NasIns" for national insect monitoring data}

\item{agg_level}{Aggregation level of data. "year_locality", "region_habitat", "region_habitat_year", "total". Default to year_locality}

\item{as_tibble}{Coerce output to class tibble}
}
\description{
obs_from_db Get insect observation data from the database
}
\examples{

dontrun{

  source("~/.rpgpass")

  connect_to_database(
     username = username,
     password = password
  )

  rm(list = c("username", "password"))


  beetles_2022 <- obs_from_db(subset_orders = "Coleoptera",
                              agg_level = "year_locality")

}



}
