% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boot_stat.R
\name{plot.boot_stat}
\alias{plot.boot_stat}
\title{Plot a boot_stat object}
\usage{
\method{plot}{boot_stat}(x, ...)
}
\arguments{
\item{x}{boot_stat_object}

\item{...}{additional parameters passed to geom_density_ridges_gradient and scale_fill_nina, e.g. bandwidth and palette}
}
\value{
A ggridges plot of the bootrapped values in a boot_stat object.
}
\description{
Plot a boot_stat object
}
\examples{


beetle_shannon_boot <- bootstrap_value(beetles,
                                       value = shannon_div,
                                       groups = c("year",
                                                  "region_name"))

plot(beetle_shannon_boot)

}
