% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_beta_part.R
\name{plot_beta_part}
\alias{plot_beta_part}
\title{plot_beta_part}
\usage{
plot_beta_part(input)
}
\arguments{
\item{input}{Tibble or dataframe of betapartition + locality distance from the combine_dist_to_comm_mat function}
}
\value{
Returns a plot of the beta diversity partition of the overall diversity
}
\description{
plot_beta_part
}
\examples{
\dontrun{
connect_to_insect_db()

seminat_trond_comm_mat <- get_community_matrix(
  trap_type = "MF",
  dataset = "NorIns",
  subset_habitat = "Semi-nat",
  subset_region = "Trøndelag",
  as_tibble = T,
  transposed_matrix = F
)

seminat_trond_dist_beta <- combine_dist_to_comm_mat(
  comm_mat = seminat_trond_comm_mat,
  region_name = "('Trøndelag')",
  habitat_type = "Semi-nat"
)

plot_beta_part(seminat_trond_dist_beta)
}

}
