% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_year_region_stats.R
\name{plot_year_region_stats}
\alias{plot_year_region_stats}
\title{plot_year_region_stats}
\usage{
plot_year_region_stats(year_region_stats)
}
\arguments{
\item{year_region_stats}{A tabular record to plot, typically from get_region_stats()}
}
\value{
A ggplot2 object
}
\description{
Conveniance function for plotting the status of the project in the staggered experimental design
}
\examples{

\dontrun{

year_region_stats <- get_year_region_stats(last_year = 2025)

plot_year_region_stats(year_region_stats)


}

}
