% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skogtype_from_db.R
\name{skogtype_from_db}
\alias{skogtype_from_db}
\title{skogtype_from_db Get the forest types from a buffer around the sampling localities}
\usage{
skogtype_from_db(
  limit = NULL,
  dataset = c("NasIns"),
  stat_grouping = c("treslag", "bonitet"),
  as_tibble = F
)
}
\arguments{
\item{limit}{Optional row limit of returned data (for testing )}

\item{dataset}{Dataset to get data from. Defaults to "NasIns"}

\item{stat_grouping}{Group data into either "treslag" or "bonitet"}

\item{as_tibble}{Return as tibble? Boolean}
}
\value{
Returns a tibble of forest data from the database.
}
\description{
skogtype_from_db Get the forest types from a buffer around the sampling localities
}
\examples{

skogtype_from_db()


}
